% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subanalysis_flags.R
\name{add_subanalysis_flags}
\alias{add_subanalysis_flags}
\title{Flag patients for subanalysis}
\usage{
add_subanalysis_flags(indication_drug_exposed, covariates, ...)
}
\arguments{
\item{indication_drug_exposed}{A tibble with columns,
c('person_id','rxcui_ingr','rxcui_ingr_name',
'group')}

\item{covariates}{A tibble with columns,
c('person_id','dob', 'gender_concept_id',
'race_concept_id', 'start_date',
'first_drug_exposure', 'final_end_date','age')}
}
\description{
\code{add_subanalysis_flags} flags patients for subanalysis
}
\details{
This function flags patients for subanalysis, based on exposure to
known approved drugs for target phenotype. Used for excluding patients
exposed to known approved drugs.

Updated: 2021-03-27
}
