% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_comorbidity_scores.R
\name{calculate_comorbidity_scores}
\alias{calculate_comorbidity_scores}
\title{Calculates Elixhauser comorbidity scores}
\usage{
calculate_comorbidity_scores(clinical_results_sub, r_i)
}
\arguments{
\item{clinical_results_sub}{A tibble}

\item{r_i}{A tibble}
}
\description{
\code{calculate_comorbidity_scores} calculates Elixhauser comorbidity scores.
}
\details{
This function calculates Elixhauser comorbidity scores.

Updated: 2021-10-02
}
