% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_process_clinical_data.R
\name{clean_process_clinical_data}
\alias{clean_process_clinical_data}
\title{Clean and process EHR data}
\usage{
clean_process_clinical_data(
  drug,
  phenotype,
  biomarker,
  indication_drugs,
  concept_id_exclusions,
  r_covariates,
  r_drugs,
  r_biomarkers
)
}
\arguments{
\item{drug}{A string, e.g., "simvastatin"}

\item{phenotype}{A string, e.g., "Hyperlipidemia"}

\item{biomarker}{A string, e.g., "LDL Cholesterol"}

\item{indication_drugs}{A tibble with columns
c('rxcui_in', 'drug_desc', 'indication'),
eg, DrugRepurposingToolKit::drugsHyperlipidemia}

\item{concept_id_exclusions}{A vector, e.g., c(2212451). Used for excluding
concept_ids which may not contain the correct
values}

\item{r_covariates}{A tibble with columns
c('person_id','dob','gender_concept_id',
'race_concept_id', drug_concept_id',
'first_drug_exposure','last_drug_exposure')}

\item{r_drugs}{A tibble with columns
c('person_id','start_date','end_date',
'drug_exposure_start_date','drug_concept_id')}

\item{r_biomarkers}{A tibble with columns
c('person_id','start_date','measurement_date',
'end_date', 'measurement_concept_id',
'value_as_number')}
}
\description{
\code{clean_process_clinical_data} cleans and processes clinical EHR data.
}
\details{
This function cleans and processes clinical data from electronic health
record (EHR) databases that have adopted the Observational Medical Outcomes
Partnership (OMOP) Common Data Model (CDM) format.

Updated: 2021-03-27
}
