% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddi_ohdsi_rxnorm}
\alias{ddi_ohdsi_rxnorm}
\title{ddi_ohdsi_rxnorm}
\format{
A data frame with 2 variables: \code{concept_id}, \code{rxcui}
}
\usage{
ddi_ohdsi_rxnorm
}
\description{
Map rxcuis to their names
}
\keyword{datasets}
