% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddi_rxcui2in}
\alias{ddi_rxcui2in}
\title{ddi_rxcui2in}
\format{
A data frame with 3 variables: \code{rxcui_ingr}, \code{rxcui} and
  \code{rxcui_ingr_name}
}
\usage{
ddi_rxcui2in
}
\description{
Map rxcuis to their rxcui ingredients
}
\keyword{datasets}
