% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddi_rxcui_names}
\alias{ddi_rxcui_names}
\title{ddi_rxcui_names}
\format{
A data frame with 3 variables: \code{drug_name}, \code{vocabulary_id} and
  \code{rxcui}
}
\usage{
ddi_rxcui_names
}
\description{
Map rxcuis to their names
}
\keyword{datasets}
