% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_observation_periods.R
\name{define_observation_periods}
\alias{define_observation_periods}
\title{Define baseline and treatment periods}
\usage{
define_observation_periods(.data, ...)
}
\arguments{
\item{.data}{A tibble}
}
\description{
\code{define_observation_periods} creates baseline and treatment period columns
}
\details{
This function defines the baseline and treatment period columns from
clinical data from electronic health record (EHR) databases that have
adopted the Observational Medical Outcomes Partnership (OMOP)
Common Data Model (CDM) format.

Baseline period = "start_date" to "first_{drug}_exposure".
Treatment period = ("first_{drug}_exposure" + 30 days) to "last_{drug}_exposure".
If treatment period > 12 months, then treatment period =
("first_{drug}_exposure" + 30 days) to "end_date".

Updated: 2021-03-27
}
