% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dr_ts_tutorial_data}
\alias{dr_ts_tutorial_data}
\title{dr_ts_tutorial_data}
\format{
List with five objects:
         c("spredixcan_phenotype_ts", "cmap_genes",
         "ilincs_output", "ilincs_simvastatin_dm4740", "pres_drugs")
}
\usage{
dr_ts_tutorial_data
}
\description{
dr_ts_tutorial_data
}
\keyword{datasets}
