% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugsHyperlipidemia}
\alias{drugsHyperlipidemia}
\title{drugsHyperlipidemia}
\format{
A data frame with 3 variables: \code{rxcui_in}, \code{drug_desc} and
  \code{indication}

A data frame with 134 rows and 3 variables:
\describe{
  \item{rxcui_in}{RxCUI ingredients}
  \item{drug_desc}{Name of active drug ingredient}
  \item{indication}{Drug indication}
  ...
}
}
\source{
MEDI-HPS (Wei W-Q, Cronin RM, Xu H, Lasko TA, Bastarache L, Denny JC.
        2013. Development and evaluation of an ensemble resource linking
        medications to their indications. J Am Med Inform Assoc 20:954–961.)
}
\usage{
drugsHyperlipidemia
}
\description{
Known lipid-lowering drugs
}
\keyword{datasets}
