% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_icd_codes.R
\name{extract_icd_codes}
\alias{extract_icd_codes}
\title{Function to get ICD data for calculating elixhauser scores}
\usage{
extract_icd_codes(
  drug_concept_id,
  biomarker_concept_id,
  table_name,
  table_type,
  elix_icds
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{table_name}{A string}

\item{table_type}{A string, either c("covariates", "drugs", "biomarkers", "icds")}

\item{elix_icds}{A string}
}
\description{
\code{extract_icd_codes} extracts ICD codes for calculating Elixhauser comorbidity scores.
}
\details{
This function extracts ICD codes for calculating Elixhauser comorbidity scores.

Updated: 2021-10-05
}
