% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drugs_to_ingredients.R
\name{map_drugs_to_ingredients}
\alias{map_drugs_to_ingredients}
\title{For drug repurposing EHR validation studies, map patient drugs to ingredients}
\usage{
map_drugs_to_ingredients(.data, ...)
}
\arguments{
\item{.data}{A tibble}
}
\description{
For drug repurposing EHR validation studies, map patient drugs to ingredients
}
\details{
Updated: 2021-03-27
}
