% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohdsi_icd_elix}
\alias{ohdsi_icd_elix}
\title{ohdsi_icd_elix}
\format{
A data frame with 5 variables: \code{concept_id}, \code{concept_name}, \code{vocabulary_id}, \code{concept_code}, \code{concept_code_strip}
}
\source{
Elixhauser, A., Steiner, C., Harris, D. R. & Coffey, R. M. Comorbidity measures for use with administrative data. Med. Care 36, 8–27 (1998)

van Walraven C, Austin PC, Jennings A, Quan H and Forster AJ. A modification of the Elixhauser comorbidity measures into a point system for hospital death using administrative data. Medical Care 2009; 47(6):626-633. DOI: 10.1097/MLR.0b013e31819432e5
}
\usage{
ohdsi_icd_elix
}
\description{
For extracting ICD codes from OMOP CDM to calculate Elixhauser comorbidity scores
}
\keyword{datasets}
