% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_biomarker.R
\name{process_biomarker}
\alias{process_biomarker}
\title{Process biomarkers data}
\usage{
process_biomarker(
  biomarker_tbl,
  biomarker_name,
  covariates_flagged,
  baseline_cols,
  inclusion_cols,
  concept_id_exclusions,
  ...
)
}
\arguments{
\item{biomarker_tbl}{A tibble with columns,
c('person_id','start_date','measurement_date',
'end_date','measurement_concept_id',
'value_as_number')}

\item{biomarker_name}{A string, e.g., "LDL cholesterol"}

\item{covariates_flagged}{A tibble with columns,
c('person_id','dob','gender_concept_id',
'race_concept_id','start_date',
'first_drug_exposure','final_end_date','age',
'drug_exposed','treatment_new_exposed')}

\item{baseline_cols}{A vector, e.g.
c('person_id', 'start_date',
'first_drug_exposure', 'final_end_date')}

\item{inclusion_cols}{A vector, e.g.,
c('person_id', 'measurement_date',
'measurement_concept_id', 'value_as_number')}

\item{concept_id_exclusions}{A vector, e.g., for LDL, c(2212451)}
}
\description{
\code{process_biomarker} sets up biomarkers for hypothesis testing
}
\details{
This function sets up biomarkers for hypothesis testing.
Updated: 2021-03-27
}
