% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_clinical_data_lmm.R
\name{process_clinical_data_lmm}
\alias{process_clinical_data_lmm}
\title{Clean and process EHR data for linear mixed model analysis}
\usage{
process_clinical_data_lmm(clinical_results, r_i)
}
\arguments{
\item{clinical_results}{A tibble}

\item{r_i}{A tibble}
}
\description{
\code{process_clinical_data_lmm} clean and process EHR data for linear mixed model analysis.
}
\details{
This function cleans and processes clinical data from electronic health
record (EHR) databases that have adopted the Observational Medical Outcomes
Partnership (OMOP) Common Data Model (CDM) format, for linear mixed model analysis.

Updated: 2021-10-02
}
