% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vw_weights}
\alias{vw_weights}
\title{vw_weights}
\format{
A data frame with 2 variables: \code{group} and \code{weight}

A data frame with 31 rows and 2 variables:
\describe{
  \item{group}{Elixhauser comorbidity group. From Elixhauser et al 1998}
  \item{weight}{Comorbidity weight. From van Walraven et al 2009}
}
}
\source{
Elixhauser, A., Steiner, C., Harris, D. R. & Coffey, R. M. Comorbidity measures for use with administrative data. Med. Care 36, 8–27 (1998)

van Walraven C, Austin PC, Jennings A, Quan H and Forster AJ. A modification of the Elixhauser comorbidity measures into a point system for hospital death using administrative data. Medical Care 2009; 47(6):626-633. DOI: 10.1097/MLR.0b013e31819432e5
}
\usage{
vw_weights
}
\description{
For calculating Elixhauser comorbidity scores
}
\keyword{datasets}
